% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmtsr-package.R
\docType{package}
\name{wmtsr-package}
\alias{wmtsr}
\alias{wmtsr-package}
\title{wmtsr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
