% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmtsr-package.R
\docType{package}
\name{wombats-package}
\alias{wombats}
\alias{wombats-package}
\title{wombats: What the Package Does (Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
